pull calculation of dates functionality into a helper
