+++ app/model/reports/CampaignPageViewsReport.scala
@@ -18,7 +18,7 @@ case class CampaignPageViewsReport(campaignId: String, seenPaths: Set[String], p
       lastData <- pageCountStats.lastOption;
       lastSeenDate <- lastData.get("date")
     ) {
-      val missingDays = CampaignPageViewsReport.calculateDatesToFetch(new DateTime(lastSeenDate).plusDays(1), DateTime.now)
+      val missingDays = DateBasedReport.calculateDatesToFetch(new DateTime(lastSeenDate).plusDays(1), DateTime.now)
       val dailyReports = missingDays.map(CampaignPageViewsReport.loadCampaignPageViewsForDay(campaign, _))
 
       val refreshed = dailyReports.foldLeft(this) { case (report: CampaignPageViewsReport, (date: DateTime, dailyViewCounts: DailyViewCounts)) =>
@@ -125,7 +125,7 @@ object CampaignPageViewsReport {
       campaign <- CampaignRepository.getCampaign(campaignId);
       startDate <- campaign.startDate
     ) yield {
-      val dailyReports = calculateDatesToFetch(startDate, DateTime.now).map{ dt =>
+      val dailyReports = DateBasedReport.calculateDatesToFetch(startDate, DateTime.now).map{ dt =>
         Thread.sleep(1000) // try to avoid rate limiting
         loadCampaignPageViewsForDay(campaign, dt)
       }
@@ -166,20 +166,4 @@ object CampaignPageViewsReport {
     }
   }
 
-  val GA_SWITCH_ON_DATE = new DateTime("2016-07-01")
-
-  def calculateDatesToFetch(startDate: DateTime, endDate: DateTime): List[DateTime] = {
-
-    val sd = if(startDate.isBefore(GA_SWITCH_ON_DATE)) GA_SWITCH_ON_DATE else startDate
-    var date = sd.withTimeAtStartOfDay()
-    val endDay = endDate.withTimeAtStartOfDay()
-    var daysInRange: List[DateTime] = Nil
-
-    while (date.isBefore(endDay)) {
-      daysInRange = daysInRange :+ date
-      date = date.plusDays(1)
-    }
-
-    daysInRange
-  }
 }
